Fix compilation errors from int->long price type widening
StockService: int listingPrice -> long to match calcListingPrice return type
StockSplitService: cast SPLIT_THRESHOLD_PRICE to int for StockSplitNotice
field (threshold is 1_000_000 which fits safely in int)

+++ backend/src/main/java/com/spotchzxk/service/StockService.java
@@ -110,7 +110,7 @@ public class StockService {
         }
         tradeEngine.evictUserCache(userId);
 
-        int listingPrice = calcListingPrice(stock.getFollowerCount());
+        long listingPrice = calcListingPrice(stock.getFollowerCount());
         stock.finalizeListing(listingPrice, 100_000L);
 
         int tier = AmmMigrationService.calcLiquidityTier(stock.getFollowerCount());
